# .kilo/worktrees/equinox-ocean/docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "JurisLab API",
  "description": "Multi-agent legal simulation and document review platform (educational).",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/users": {
   "get": {
    "tags": [
     "users"
    ],
    "summary": "List Users",
    "operationId": "list_users_api_v1_users_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response List Users Api V1 Users Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/users/{user_id}": {
   "get": {
    "tags": [
     "users"
    ],
    "summary": "Get User",
    "operationId": "get_user_api_v1_users__user_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "User Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get User Api V1 Users  User Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/users/{user_id}/roles": {
   "put": {
    "tags": [
     "users"
    ],
    "summary": "Assign Roles",
    "operationId": "assign_roles_api_v1_users__user_id__roles_put",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "User Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RoleAssign"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Assign Roles Api V1 Users  User Id  Roles Put"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/users/invite": {
   "post": {
    "tags": [
     "users"
    ],
    "summary": "Invite User",
    "operationId": "invite_user_api_v1_users_invite_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "email",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Email"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Invite User Api V1 Users Invite Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "RoleAssign": {
    "properties": {
     "organization_id": {
      "type": "string",
      "title": "Organization Id"
     },
     "platform_role": {
      "type": "string",
      "title": "Platform Role",
      "default": "member"
     },
     "module_roles": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Module Roles",
      "default": []
     }
    },
    "type": "object",
    "required": [
     "organization_id"
    ],
    "title": "RoleAssign"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 },
...
}